requirements.txt
openai>=1.40.0
pandas>=2.2.2
httpx>=0.27.0
python-dotenv>=1.0.1
tqdm>=4.66.4
